chore: Move to use a src directory
