+++ scripts/setup_for_wrf.py
@@ -1,10 +1,8 @@
 import json
 import datetime
 import pytz
-import io
 import re
 import os
-import errno    
 import argparse
 import math
 import f90nml
@@ -13,12 +11,11 @@ import subprocess
 import glob
 import sys
 import pdb
-import time
 import resource
 import copy
 import stat
 import netCDF4
-from downloadFNLanalyses import download_gdas_fnl_data
+from setup_runs.wrf.fetch_fnl import download_gdas_fnl_data
 
 ## get command line arguments
 parser = argparse.ArgumentParser()
